only create (shareable) tables if they not exist, add is_top_concept column if not exist

+++ lumen/database/migrations/2016_11_09_000001_create_th_broaders_table.php
@@ -13,15 +13,17 @@ class CreateThBroadersTable extends Migration
      */
     public function up()
     {
-        Schema::create('th_broaders', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('broader_id')->unsigned();
-            $table->integer('narrower_id')->unsigned();
-            $table->timestamps();
+        if(!Schema::hasTable('th_broaders')) {
+            Schema::create('th_broaders', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('broader_id')->unsigned();
+                $table->integer('narrower_id')->unsigned();
+                $table->timestamps();
 
-            $table->foreign('broader_id')->references('id')->on('th_concept')->onDelete('cascade');
-            $table->foreign('narrower_id')->references('id')->on('th_concept')->onDelete('cascade');
-        });
+                $table->foreign('broader_id')->references('id')->on('th_concept')->onDelete('cascade');
+                $table->foreign('narrower_id')->references('id')->on('th_concept')->onDelete('cascade');
+            });
+        }
     }
 
     /**
