+++ lumen/database/migrations/2016_11_09_000001_create_th_concept_label_table.php
@@ -13,18 +13,20 @@ class CreateThConceptLabelTable extends Migration
      */
     public function up()
     {
-        Schema::create('th_concept_label', function (Blueprint $table) {
-            $table->increments('id');
-            $table->text('lasteditor');
-            $table->text('label');
-            $table->integer('concept_id')->unsigned();
-            $table->integer('language_id')->unsigned();
-            $table->integer('concept_label_type')->default(1);
-            $table->timestamps();
+        if(!Schema::hasTable('th_concept_label')) {
+            Schema::create('th_concept_label', function (Blueprint $table) {
+                $table->increments('id');
+                $table->text('lasteditor');
+                $table->text('label');
+                $table->integer('concept_id')->unsigned();
+                $table->integer('language_id')->unsigned();
+                $table->integer('concept_label_type')->default(1);
+                $table->timestamps();
 
-            $table->foreign('concept_id')->references('id')->on('th_concept')->onDelete('cascade');
-            $table->foreign('language_id')->references('id')->on('th_language')->onDelete('cascade');
-        });
+                $table->foreign('concept_id')->references('id')->on('th_concept')->onDelete('cascade');
+                $table->foreign('language_id')->references('id')->on('th_language')->onDelete('cascade');
+            });
+        }
     }
 
     /**
